+++ mongo-utils/pom.xml
@@ -5,7 +5,7 @@
     <parent>
         <artifactId>module-mongo</artifactId>
         <groupId>org.wso2.ei</groupId>
-        <version>0.9.0</version>
+        <version>0.9.1</version>
     </parent>
 
     <modelVersion>4.0.0</modelVersion>
@@ -17,38 +17,16 @@
             <artifactId>ballerina-runtime</artifactId>
         </dependency>
         <dependency>
-            <groupId>org.ballerinalang.platform</groupId>
-            <artifactId>playground-parser</artifactId>
-            <version>0.8-SNAPSHOT</version>
-        </dependency>
-        <dependency>
-            <groupId>org.ballerinalang.platform</groupId>
-            <artifactId>playground-parser</artifactId>
-            <version>0.8-SNAPSHOT</version>
-        </dependency>
-        <dependency>
             <groupId>org.mongodb</groupId>
             <artifactId>mongo-java-driver</artifactId>
-            <version>3.8.2</version>
-        </dependency>
-        <dependency>
-            <groupId>org.mongodb</groupId>
-            <artifactId>mongo-java-driver</artifactId>
-            <version>3.8.2</version>
         </dependency>
         <dependency>
             <groupId>org.mongodb</groupId>
             <artifactId>mongodb-driver-core</artifactId>
-            <version>3.8.2</version>
         </dependency>
         <dependency>
             <groupId>org.mongodb</groupId>
             <artifactId>bson</artifactId>
-            <version>2.3</version>
         </dependency>
     </dependencies>
-
-
 </project>
-
-
