+++ mongo-utils/src/main/java/org/wso2/mongo/BallerinaMongoDbException.java
@@ -1,0 +1,16 @@
+package org.wso2.mongo;
+
+import org.ballerinalang.jvm.util.exceptions.BallerinaException;
+
+/**
+ * Exception class for mongo
+ */
+public class BallerinaMongoDbException extends BallerinaException {
+    public BallerinaMongoDbException(String message) {
+        super(message);
+    }
+
+    public BallerinaMongoDbException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
