+++ mongodb/Ballerina.toml
@@ -9,7 +9,7 @@ target = "java8"
 
 [[platform.libraries]]
 module = "mongodb"
-path = "./utils/target/mongo-utils-0.9.0.jar"
+path = "../mongo-utils/target/mongo-utils-0.9.1.jar"
 artafactId = "mongo-utils"
 version = "0.9.1"
 groupId = "org.wso2.ei"
