+++ pom.xml
@@ -27,11 +27,12 @@
     <groupId>org.wso2.ei</groupId>
     <artifactId>module-mongo</artifactId>
     <packaging>pom</packaging>
-    <version>0.9.0</version>
+    <version>0.9.1</version>
     <name>Ballerina MongoDB Module</name>
 
     <modules>
-        <module>utils</module>
+        <module>mongo-utils</module>
+        <module>mongodb</module>
     </modules>
 
     <dependencyManagement>
@@ -54,11 +55,6 @@
                 <version>${ballerina.version}</version>
                 <scope>compile</scope>
             </dependency>
-            <!--<dependency>-->
-            <!--<groupId>javax.jms</groupId>-->
-            <!--<artifactId>javax.jms-api</artifactId>-->
-            <!--<version>${jms.version}</version>-->
-            <!--</dependency>-->
             <dependency>
                 <groupId>org.slf4j</groupId>
                 <artifactId>slf4j-simple</artifactId>
@@ -75,44 +71,32 @@
                 <artifactId>testng</artifactId>
                 <scope>test</scope>
             </dependency>
-            <!-- https://mvnrepository.com/artifact/org.mongodb/mongo-java-driver -->
             <dependency>
                 <groupId>org.mongodb</groupId>
                 <artifactId>mongo-java-driver</artifactId>
-                <version>2.1</version>
+                <version>3.8.2</version>
             </dependency>
             <dependency>
-            <groupId>org.mongodb</groupId>
-            <artifactId>mongo-java-driver</artifactId>
-            <version>3.8.2</version>
-        </dependency>
-        <dependency>
-            <groupId>org.mongodb</groupId>
-            <artifactId>mongo-java-driver</artifactId>
-            <version>3.8.2</version>
-        </dependency>
-        <dependency>
-            <groupId>org.mongodb</groupId>
-            <artifactId>mongodb-driver-core</artifactId>
-            <version>3.8.2</version>
-        </dependency>
-        <dependency>
-            <groupId>org.mongodb</groupId>
-            <artifactId>bson</artifactId>
-            <version>2.3</version>
-        </dependency>
+                <groupId>org.mongodb</groupId>
+                <artifactId>mongodb-driver-core</artifactId>
+                <version>3.8.2</version>
+            </dependency>
+            <dependency>
+                <groupId>org.mongodb</groupId>
+                <artifactId>bson</artifactId>
+                <version>3.8.2</version>
+            </dependency>
         </dependencies>
     </dependencyManagement>
 
     <properties>
         <project-home>${project.basedir}</project-home>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <!--ballerina.version>1.0.0-alpha3</ballerina.version-->
+        <ballerina.version>1.0.1</ballerina.version>
         <ballerina.version>1.0.0-beta</ballerina.version>
         <maven.checkstyleplugin.excludes>**/generated/**</maven.checkstyleplugin.excludes>
         <maven.spotbugsplugin.exclude.file>spotbugs-exclude.xml</maven.spotbugsplugin.exclude.file>
         <slf4j.version>1.7.21</slf4j.version>
-        <!--<jms.version>2.0.1</jms.version>-->
     </properties>
     <profiles>
         <profile>
@@ -136,7 +120,6 @@
                             <executable>ballerina</executable>
                             <arguments>
                                 <argument>compile</argument>
-                                <!--                                <argument>jms</argument>-->
                             </arguments>
                         </configuration>
                     </plugin>
